requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.4
mysqlclient==2.2.0
django-redis==5.4.0
pycryptodomex==3.19.0
djangorestframework-simplejwt==5.3.0
celery==5.3.6
django-celery-results==2.5.1
django-celery-beat==2.5.0
flower==2.0.1
python-daemon==3.0.1
gunicorn==21.2.0
django-proxy==1.2.2
psutil==5.9.6
uvicorn==0.24.0.post1
daphne==4.0.0
channels==4.0.0
channels-redis==4.1.0
django-simple-captcha==0.5.20
user-agents==2.2.0
aiofiles==23.2.1
websockets==12.0
requests==2.31.0
datclass==0.2.26